fix: Resolve 8 CI type checking errors
- Remove unused type: ignore[import-not-found] from database.py (asyncpg, aiosqlite)
- Remove unused type: ignore[import-not-found] from ppo_trainer.py (stable_baselines3)
- Add asyncpg, aiosqlite, stable_baselines3 to ignore_missing_imports in pyproject.toml
- Wrap numpy float64 results in float() in epidemic_model.py and biometric.py
- Replace deprecated np.trapz with np.trapezoid in metrics.py
- Rename is_anomaly to anomaly_mask in lavad.py to avoid bool/ndarray type conflict

+++ src/omni_mercury_engine/models/biometric.py
@@ -279,7 +279,7 @@ class BiometricAnomalyModel:
             if TORCH_AVAILABLE and isinstance(features, torch.Tensor):
                 features = features.detach().cpu().numpy()
 
-            anomaly_score = np.mean(np.abs(features - 0.5))
+            anomaly_score = float(np.mean(np.abs(features - 0.5)))
 
             return {
                 "model_type": "biometric",
